+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Renamed `seriesColors.upToEight` to `seriesColors.limited`.
+- Removed `Olive` & `Orange` colors from `Light` theme.
 
 ## [13.4.1] - 2024-06-20